Получение объекта по ID, и различные оптимизации
