--- app/Dockerfile
@@ -1,11 +1,0 @@
-FROM python:3.9
-
-WORKDIR /app
-
-COPY requirements.txt .
-
-RUN pip install --upgrade pip && pip install -r requirements.txt --no-cache-dir
-
-COPY . .
-
-CMD uvicorn main:app --reload
