+++ app/auth/crud.py
@@ -1,8 +1,7 @@
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.models.models import User
-
+from .models import User
 from .security import get_password_hash
