+++ app/products/crud.py
@@ -11,3 +11,10 @@ async def get_all_data(db: AsyncSession, model):
 
     result = await db.execute(select(model).limit(18))
     return result.scalars().all()
+
+
+async def get_data_by_id(db: AsyncSession, model, id):
+    """Получение объекта из выбранной модели по ID."""
+
+    result = await db.execute(select(model).filter(model.id == id))
+    return result.scalars().one_or_none()
