+++ docker-compose.yml
@@ -1,0 +1,28 @@
+version: '3.8'
+
+services:
+
+  db:
+    container_name: postgresql
+    image: postgres:13.0-alpine
+    restart: always
+    ports:
+      - "5432:5432"
+    volumes:
+      - database:/var/lib/postgresql/data/
+    env_file:
+      - ./.env
+
+  backend:
+    container_name: backend
+    build: .
+    restart: always
+    ports:
+      - "8000:80"
+    depends_on:
+      - db
+    env_file:
+      - ./.env
+
+volumes:
+  database:
