+++ requirements.txt
@@ -42,6 +42,7 @@ PyYAML==6.0.1
 rsa==4.9
 six==1.16.0
 sniffio==1.3.0
+sqladmin==0.16.0
 SQLAlchemy==2.0.23
 starlette==0.27.0
 typing_extensions==4.8.0
@@ -49,3 +50,4 @@ ujson==5.8.0
 uvicorn==0.24.0.post1
 watchfiles==0.21.0
 websockets==12.0
+WTForms==3.1.1
